requirements.txt
streamlit
python-dotenv
PyPDF2
gTTS
openai
langchain
langchain_community
langchain-groq
huggingface_hub
torch
sentence-transformers
faiss-cpu
